break out network overview and home_search_bar
